+++ rust/operator-binary/src/airflow_controller.rs
@@ -1,6 +1,6 @@
 //! Ensures that `Pod`s are configured and running for each [`AirflowCluster`]
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::{BTreeMap, BTreeSet, HashMap},
     io::Write,
     str::FromStr,
     sync::Arc,
@@ -13,8 +13,11 @@ use product_config::{
 };
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_airflow_crd::{
-    authentication::AirflowAuthenticationConfigResolved, build_recommended_labels,
-    git_sync::GitSync, AirflowCluster, AirflowClusterStatus, AirflowConfig, AirflowConfigFragment,
+    authentication::AirflowAuthenticationClassResolved, git_sync::GitSync,
+};
+use stackable_airflow_crd::{
+    authentication::AirflowClientAuthenticationDetailsResolved, build_recommended_labels,
+    AirflowCluster, AirflowClusterStatus, AirflowConfig, AirflowConfigFragment,
     AirflowConfigOptions, AirflowExecutor, AirflowRole, Container, ExecutorConfig,
     ExecutorConfigFragment, AIRFLOW_CONFIG_FILENAME, AIRFLOW_UID, APP_NAME, CONFIG_PATH,
     GIT_CONTENT, GIT_ROOT, GIT_SYNC_NAME, LOG_CONFIG_DIR, OPERATOR_NAME, STACKABLE_LOG_DIR,
@@ -32,7 +35,7 @@ use stackable_operator::{
     },
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
     commons::{
-        authentication::{ldap, AuthenticationClass, AuthenticationClassProvider},
+        authentication::{ldap, AuthenticationClass},
         product_image_selection::ResolvedProductImage,
         rbac::build_rbac_resources,
     },
@@ -74,6 +77,7 @@ use stackable_operator::{
     time::Duration,
     utils::COMMON_BASH_TRAP_FUNCTIONS,
 };
+
 use strum::{EnumDiscriminants, IntoEnumIterator, IntoStaticStr};
 
 use crate::{
@@ -302,6 +306,16 @@ pub enum Error {
         source: builder::pod::container::Error,
     },
 
+    #[snafu(display("failed to add LDAP Volumes and VolumeMounts"))]
+    AddLdapVolumesAndVolumeMounts {
+        source: stackable_operator::commons::authentication::ldap::Error,
+    },
+
+    #[snafu(display("failed to add TLS Volumes and VolumeMounts"))]
+    AddTlsVolumesAndVolumeMounts {
+        source: stackable_operator::commons::tls_verification::TlsClientDetailsError,
+    },
+
     #[snafu(display("AirflowCluster object is invalid"))]
     InvalidAirflowCluster {
         source: error_boundary::InvalidObject,
@@ -337,13 +351,12 @@ pub async fn reconcile_airflow(
     let cluster_operation_cond_builder =
         ClusterOperationsConditionBuilder::new(&airflow.spec.cluster_operation);
 
-    let authentication_config = airflow
-        .spec
-        .cluster_config
-        .authentication
-        .resolve(client)
-        .await
-        .context(InvalidAuthenticationConfigSnafu)?;
+    let authentication_config = AirflowClientAuthenticationDetailsResolved::from(
+        &airflow.spec.cluster_config.authentication,
+        client,
+    )
+    .await
+    .context(InvalidAuthenticationConfigSnafu)?;
 
     let mut roles = HashMap::new();
 
@@ -474,7 +487,7 @@ pub async fn reconcile_airflow(
                 &airflow_role,
                 &rolegroup,
                 rolegroup_config,
-                authentication_config.as_ref(),
+                &authentication_config,
                 &rbac_sa.name_unchecked(),
                 &merged_airflow_config,
                 airflow_executor,
@@ -494,7 +507,7 @@ pub async fn reconcile_airflow(
                 &resolved_product_image,
                 &rolegroup,
                 rolegroup_config,
-                authentication_config.as_ref(),
+                &authentication_config,
                 &merged_airflow_config.logging,
                 vector_aggregator_address.as_deref(),
                 &Container::Airflow,
@@ -543,7 +556,7 @@ async fn build_executor_template(
     airflow: &AirflowCluster,
     common_config: &CommonConfiguration<ExecutorConfigFragment>,
     resolved_product_image: &ResolvedProductImage,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
     vector_aggregator_address: &Option<String>,
     cluster_resources: &mut ClusterResources,
     client: &stackable_operator::client::Client,
@@ -557,6 +570,7 @@ async fn build_executor_template(
         role: "executor".into(),
         role_group: "kubernetes".into(),
     };
+
     let rg_configmap = build_rolegroup_config_map(
         airflow,
         resolved_product_image,
@@ -659,7 +673,7 @@ fn build_rolegroup_config_map(
     resolved_product_image: &ResolvedProductImage,
     rolegroup: &RoleGroupRef<AirflowCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
     logging: &Logging<Container>,
     vector_aggregator_address: Option<&str>,
     container: &Container,
@@ -821,7 +835,7 @@ fn build_server_rolegroup_statefulset(
     airflow_role: &AirflowRole,
     rolegroup_ref: &RoleGroupRef<AirflowCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
     sa_name: &str,
     merged_airflow_config: &AirflowConfig,
     executor: &AirflowExecutor,
@@ -888,6 +902,7 @@ fn build_server_rolegroup_statefulset(
         airflow_role,
         rolegroup_config,
         executor,
+        authentication_config,
     ));
 
     let volume_mounts = airflow.volume_mounts();
@@ -1098,7 +1113,7 @@ fn build_logging_container(
 fn build_executor_template_config_map(
     airflow: &AirflowCluster,
     resolved_product_image: &ResolvedProductImage,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
     sa_name: &str,
     merged_executor_config: &ExecutorConfig,
     env_overrides: &HashMap<String, String>,
@@ -1138,12 +1153,12 @@ fn build_executor_template_config_map(
     let mut airflow_container =
         ContainerBuilder::new(&Container::Base.to_string()).context(InvalidContainerNameSnafu)?;
 
+    // Works too, had been changed
     add_authentication_volumes_and_volume_mounts(
         authentication_config,
         &mut airflow_container,
         &mut pb,
     )?;
-
     airflow_container
         .image_from_product_image(resolved_product_image)
         .resources(merged_executor_config.resources.clone().into())
@@ -1283,28 +1298,40 @@ pub fn error_policy(
         _ => Action::requeue(*Duration::from_secs(10)),
     }
 }
+// I want to add secret volumes right here
 
 fn add_authentication_volumes_and_volume_mounts(
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
     cb: &mut ContainerBuilder,
     pb: &mut PodBuilder,
 ) -> Result<()> {
-    // TODO: Currently there can be only one AuthenticationClass due to FlaskAppBuilder restrictions.
-    //    Needs adaptation once FAB and airflow support multiple auth methods.
-    // The checks for max one AuthenticationClass and the provider are done in crd/src/authentication.rs
-    for config in authentication_config {
-        if let Some(auth_class) = &config.authentication_class {
-            match &auth_class.spec.provider {
-                AuthenticationClassProvider::Ldap(ldap) => {
-                    ldap.add_volumes_and_mounts(pb, vec![cb])
-                        .context(VolumeAndMountsSnafu)?;
-                }
-                AuthenticationClassProvider::Tls(_)
-                | AuthenticationClassProvider::Oidc(_)
-                | AuthenticationClassProvider::Static(_)
-                | AuthenticationClassProvider::Kerberos(_) => {}
+    // Different authentication entries can reference the same secret
+    // class or TLS certificate. It must be ensured that the volumes
+    // and volume mounts are only added once in such a case.
+
+    let mut ldap_authentication_providers = BTreeSet::new();
+    let mut tls_client_credentials = BTreeSet::new();
+
+    for auth_class_resolved in &authentication_config.authentication_classes_resolved {
+        match auth_class_resolved {
+            AirflowAuthenticationClassResolved::Ldap { provider } => {
+                ldap_authentication_providers.insert(provider);
+            }
+            AirflowAuthenticationClassResolved::Oidc { provider, .. } => {
+                tls_client_credentials.insert(&provider.tls);
             }
         }
+    }
+
+    for provider in ldap_authentication_providers {
+        provider
+            .add_volumes_and_mounts(pb, vec![cb])
+            .context(AddLdapVolumesAndVolumeMountsSnafu)?;
+    }
+
+    for tls in tls_client_credentials {
+        tls.add_volumes_and_mounts(pb, vec![cb])
+            .context(AddTlsVolumesAndVolumeMountsSnafu)?;
     }
     Ok(())
 }