+++ rust/operator-binary/src/config.rs
@@ -1,11 +1,13 @@
+use indoc::formatdoc;
 use snafu::{ResultExt, Snafu};
 use stackable_airflow_crd::{
-    authentication::AirflowAuthenticationConfigResolved, authentication::FlaskRolesSyncMoment,
+    authentication::{
+        AirflowAuthenticationClassResolved, AirflowClientAuthenticationDetailsResolved,
+        FlaskRolesSyncMoment, DEFAULT_OIDC_PROVIDER,
+    },
     AirflowConfigOptions,
 };
-use stackable_operator::commons::authentication::{
-    ldap::AuthenticationProvider, AuthenticationClassProvider,
-};
+use stackable_operator::commons::authentication::{ldap::AuthenticationProvider, oidc};
 use stackable_operator::commons::tls_verification::TlsVerification;
 use std::collections::BTreeMap;
 
@@ -28,7 +30,7 @@ pub enum Error {
 
 pub fn add_airflow_config(
     config: &mut BTreeMap<String, String>,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
+    authentication_config: &AirflowClientAuthenticationDetailsResolved,
 ) -> Result<()> {
     if !config.contains_key(&*AirflowConfigOptions::AuthType.to_string()) {
         config.insert(
@@ -45,31 +47,53 @@ pub fn add_airflow_config(
 
 fn append_authentication_config(
     config: &mut BTreeMap<String, String>,
-    authentication_config: &Vec<AirflowAuthenticationConfigResolved>,
-) -> Result<()> {
-    // TODO: we make sure in crd/src/authentication.rs that currently there is only one
-    //    AuthenticationClass provided. If the FlaskAppBuilder ever supports this we have
-    //    to adapt the config here accordingly
-    for auth_config in authentication_config {
-        if let Some(auth_class) = &auth_config.authentication_class {
-            if let AuthenticationClassProvider::Ldap(ldap) = &auth_class.spec.provider {
-                append_ldap_config(config, ldap)?;
+    auth_config: &AirflowClientAuthenticationDetailsResolved,
+) -> Result<(), Error> {
+    let ldap_providers = auth_config
+        .authentication_classes_resolved
+        .iter()
+        .filter_map(|auth_class| {
+            if let AirflowAuthenticationClassResolved::Ldap { provider } = auth_class {
+                Some(provider)
+            } else {
+                None
             }
-        }
+        })
+        .collect::<Vec<_>>();
 
-        config.insert(
-            AirflowConfigOptions::AuthUserRegistration.to_string(),
-            auth_config.user_registration.to_string(),
-        );
-        config.insert(
-            AirflowConfigOptions::AuthUserRegistrationRole.to_string(),
-            auth_config.user_registration_role.to_string(),
-        );
-        config.insert(
-            AirflowConfigOptions::AuthRolesSyncAtLogin.to_string(),
-            (auth_config.sync_roles_at == FlaskRolesSyncMoment::Login).to_string(),
-        );
+    let oidc_providers = auth_config
+        .authentication_classes_resolved
+        .iter()
+        .filter_map(|auth_class| {
+            if let AirflowAuthenticationClassResolved::Oidc { provider, oidc } = auth_class {
+                Some((provider, oidc))
+            } else {
+                None
+            }
+        })
+        .collect::<Vec<_>>();
+
+    if let Some(ldap_provider) = ldap_providers.first() {
+        append_ldap_config(config, ldap_provider)?;
     }
+
+    if !oidc_providers.is_empty() {
+        append_oidc_config(config, &oidc_providers);
+    }
+
+    config.insert(
+        AirflowConfigOptions::AuthUserRegistration.to_string(),
+        auth_config.user_registration.to_string(),
+    );
+    config.insert(
+        AirflowConfigOptions::AuthUserRegistrationRole.to_string(),
+        auth_config.user_registration_role.to_string(),
+    );
+    config.insert(
+        AirflowConfigOptions::AuthRolesSyncAtLogin.to_string(),
+        (auth_config.sync_roles_at == FlaskRolesSyncMoment::Login).to_string(),
+    );
+
     Ok(())
 }
 
@@ -162,68 +186,254 @@ fn append_ldap_config(
     Ok(())
 }
 
+fn append_oidc_config(
+    config: &mut BTreeMap<String, String>,
+    providers: &[(
+        &oidc::AuthenticationProvider,
+        &oidc::ClientAuthenticationOptions<()>,
+    )],
+) {
+    // Debatable: AUTH_OAUTH or AUTH_OID
+    // Additionally can be set via config
+    config.insert(
+        AirflowConfigOptions::AuthType.to_string(),
+        "AUTH_OAUTH".into(),
+    );
+
+    let mut oauth_providers_config = Vec::new();
+
+    for (oidc, client_options) in providers {
+        let (env_client_id, env_client_secret) =
+            oidc::AuthenticationProvider::client_credentials_env_names(
+                &client_options.client_credentials_secret_ref,
+            );
+        let mut scopes = oidc.scopes.clone();
+        scopes.extend_from_slice(&client_options.extra_scopes);
+
+        let oidc_provider = oidc
+            .provider_hint
+            .as_ref()
+            .unwrap_or(&DEFAULT_OIDC_PROVIDER);
+
+        let oauth_providers_config_entry = match oidc_provider {
+            oidc::IdentityProviderHint::Keycloak => {
+                formatdoc!(
+                    "
+                      {{ 'name': 'keycloak',
+                        'icon': 'fa-key',
+                        'token_key': 'access_token',
+                        'remote_app': {{
+                          'client_id': os.environ.get('{env_client_id}'),
+                          'client_secret': os.environ.get('{env_client_secret}'),
+                          'client_kwargs': {{
+                            'scope': '{scopes}'
+                          }},
+                          'api_base_url': '{url}/protocol/',
+                          'server_metadata_url': '{url}/.well-known/openid-configuration',
+                        }},
+                      }}",
+                    url = oidc.endpoint_url().unwrap(),
+                    scopes = scopes.join(" "),
+                )
+            }
+        };
+
+        oauth_providers_config.push(oauth_providers_config_entry);
+    }
+
+    config.insert(
+        AirflowConfigOptions::OauthProviders.to_string(),
+        formatdoc!(
+            "[
+             {joined_oauth_providers_config}
+             ]
+             ",
+            joined_oauth_providers_config = oauth_providers_config.join(",\n")
+        ),
+    );
+}
+
 #[cfg(test)]
 mod tests {
     use crate::config::add_airflow_config;
+    use indoc::indoc;
     use stackable_airflow_crd::authentication::{
-        default_sync_roles_at, default_user_registration, AirflowAuthenticationConfigResolved,
+        default_sync_roles_at, default_user_registration, AirflowAuthenticationClassResolved,
+        AirflowClientAuthenticationDetailsResolved, FlaskRolesSyncMoment,
     };
-    use stackable_airflow_crd::AirflowConfigOptions;
-    use stackable_operator::commons::authentication::AuthenticationClass;
+    use stackable_operator::commons::authentication::{ldap, oidc};
     use std::collections::BTreeMap;
 
     #[test]
-    fn test_no_ldap() {
+    fn test_auth_db_config() {
+        let authentication_config = AirflowClientAuthenticationDetailsResolved {
+            authentication_classes_resolved: vec![],
+            user_registration: true,
+            user_registration_role: "User".to_string(),
+            sync_roles_at: FlaskRolesSyncMoment::Registration,
+        };
+
         let mut result = BTreeMap::new();
-        add_airflow_config(&mut result, &vec![]).expect("Ok");
+        add_airflow_config(&mut result, &authentication_config).expect("Ok");
+
         assert_eq!(
-            BTreeMap::from([("AUTH_TYPE".into(), "AUTH_DB".into())]),
+            BTreeMap::from([
+                ("AUTH_ROLES_SYNC_AT_LOGIN".into(), "false".into()),
+                ("AUTH_TYPE".into(), "AUTH_DB".into()),
+                ("AUTH_USER_REGISTRATION".into(), "true".into()),
+                ("AUTH_USER_REGISTRATION_ROLE".into(), "User".into())
+            ]),
             result
         );
     }
 
     #[test]
-    fn test_ldap() {
-        let authentication_class = "
-            apiVersion: authentication.stackable.tech/v1alpha1
-            kind: AuthenticationClass
-            metadata:
-              name: airflow-with-ldap-server-veri-tls-ldap
-            spec:
-              provider:
-                ldap:
-                  hostname: openldap.default.svc.cluster.local
-                  port: 636
-                  searchBase: ou=users,dc=example,dc=org
-                  ldapFieldNames:
-                    uid: uid
-                  bindCredentials:
-                    secretClass: airflow-with-ldap-server-veri-tls-ldap-bind
-                  tls:
-                    verification:
-                      server:
-                        caCert:
-                          secretClass: openldap-tls
-          ";
-        let deserializer = serde_yaml::Deserializer::from_str(authentication_class);
-        let authentication_class: AuthenticationClass =
+    fn test_ldap_config() {
+        let ldap_provider_yaml = r#"
+            hostname: openldap.default.svc.cluster.local
+            port: 636
+            searchBase: ou=users,dc=example,dc=org
+            ldapFieldNames:
+              uid: uid
+            bindCredentials:
+              secretClass: airflow-with-ldap-server-veri-tls-ldap-bind
+            tls:
+              verification:
+                server:
+                  caCert:
+                    secretClass: openldap-tls
+        "#;
+        let deserializer = serde_yaml::Deserializer::from_str(ldap_provider_yaml);
+        let ldap_provider: ldap::AuthenticationProvider =
             serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
 
-        let resolved_config = AirflowAuthenticationConfigResolved {
-            authentication_class: Some(authentication_class),
+        let authentication_config = AirflowClientAuthenticationDetailsResolved {
+            authentication_classes_resolved: vec![AirflowAuthenticationClassResolved::Ldap {
+                provider: ldap_provider,
+            }],
+            user_registration: true,
+            user_registration_role: "Admin".to_string(),
+            sync_roles_at: FlaskRolesSyncMoment::Registration,
+        };
+
+        let mut result = BTreeMap::new();
+        add_airflow_config(&mut result, &authentication_config).expect("Ok");
+
+        assert_eq!(BTreeMap::from([
+            ("AUTH_LDAP_ALLOW_SELF_SIGNED".into(), "false".into()),
+            ("AUTH_LDAP_BIND_PASSWORD".into(), "open('/stackable/secrets/airflow-with-ldap-server-veri-tls-ldap-bind/password').read()".into()),
+            ("AUTH_LDAP_BIND_USER".into(), "open('/stackable/secrets/airflow-with-ldap-server-veri-tls-ldap-bind/user').read()".into()),
+            ("AUTH_LDAP_FIRSTNAME_FIELD".into(), "givenName".into()),
+            ("AUTH_LDAP_GROUP_FIELD".into(), "memberof".into()),
+            ("AUTH_LDAP_LASTNAME_FIELD".into(), "sn".into()),
+            ("AUTH_LDAP_SEARCH".into(), "ou=users,dc=example,dc=org".into()),
+            ("AUTH_LDAP_SEARCH_FILTER".into(), "".into()),
+            ("AUTH_LDAP_SERVER".into(), "ldaps://openldap.default.svc.cluster.local:636".into()),
+            ("AUTH_LDAP_TLS_CACERTFILE".into(), "/stackable/secrets/openldap-tls/ca.crt".into()),
+            ("AUTH_LDAP_TLS_DEMAND".into(), "true".into()),
+            ("AUTH_LDAP_UID_FIELD".into(), "uid".into()),
+            ("AUTH_ROLES_SYNC_AT_LOGIN".into(), "false".into()),
+            ("AUTH_TYPE".into(), "AUTH_LDAP".into()),
+            ("AUTH_USER_REGISTRATION".into(), "true".into()),
+            ("AUTH_USER_REGISTRATION_ROLE".into(), "Admin".into())
+        ]), result);
+    }
+
+    #[test]
+    fn test_oidc_config() {
+        let oidc_provider_yaml1 = r#"
+            hostname: my.keycloak1.server
+            port: 12345
+            rootPath: my-root-path
+            tls:
+              verification:
+                server:
+                  caCert:
+                    secretClass: keycloak-ca-cert
+            principalClaim: preferred_username
+            scopes:
+              - openid
+              - email
+              - profile
+            provider_hint: Keycloak
+        "#;
+        let deserializer = serde_yaml::Deserializer::from_str(oidc_provider_yaml1);
+        let oidc_provider1: oidc::AuthenticationProvider =
+            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
+
+        let oidc_provider_yaml2 = r#"
+            hostname: my.keycloak2.server
+            principalClaim: preferred_username
+            scopes:
+              - openid
+            provider_hint: Keycloak
+        "#;
+        let deserializer = serde_yaml::Deserializer::from_str(oidc_provider_yaml2);
+        let oidc_provider2: oidc::AuthenticationProvider =
+            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
+
+        let authentication_config = AirflowClientAuthenticationDetailsResolved {
+            authentication_classes_resolved: vec![
+                AirflowAuthenticationClassResolved::Oidc {
+                    provider: oidc_provider1,
+                    oidc: oidc::ClientAuthenticationOptions {
+                        client_credentials_secret_ref: "test-client-secret1".to_string(),
+                        extra_scopes: vec!["roles".to_string()],
+                        product_specific_fields: (),
+                    },
+                },
+                AirflowAuthenticationClassResolved::Oidc {
+                    provider: oidc_provider2,
+                    oidc: oidc::ClientAuthenticationOptions {
+                        client_credentials_secret_ref: "test-client-secret2".to_string(),
+                        extra_scopes: vec![],
+                        product_specific_fields: (),
+                    },
+                },
+            ],
             user_registration: default_user_registration(),
             user_registration_role: "Admin".to_string(),
             sync_roles_at: default_sync_roles_at(),
         };
 
         let mut result = BTreeMap::new();
-        add_airflow_config(&mut result, &vec![resolved_config]).expect("Ok");
+        add_airflow_config(&mut result, &authentication_config).expect("Ok");
 
-        assert_eq!(
-            "AUTH_LDAP",
-            result
-                .get(&AirflowConfigOptions::AuthType.to_string())
-                .unwrap()
-        );
+        assert_eq!(BTreeMap::from([
+            ("AUTH_ROLES_SYNC_AT_LOGIN".into(), "false".into()),
+            ("AUTH_TYPE".into(), "AUTH_OAUTH".into()),
+            ("AUTH_USER_REGISTRATION".into(), "true".into()),
+            ("AUTH_USER_REGISTRATION_ROLE".into(), "Admin".into()),
+            ("OAUTH_PROVIDERS".into(), indoc!("
+              [
+              { 'name': 'keycloak',
+                'icon': 'fa-key',
+                'token_key': 'access_token',
+                'remote_app': {
+                  'client_id': os.environ.get('OIDC_A96BCC4FA49835D2_CLIENT_ID'),
+                  'client_secret': os.environ.get('OIDC_A96BCC4FA49835D2_CLIENT_SECRET'),
+                  'client_kwargs': {
+                    'scope': 'openid email profile roles'
+                  },
+                  'api_base_url': 'https://my.keycloak1.server:12345/my-root-path/protocol/',
+                  'server_metadata_url': 'https://my.keycloak1.server:12345/my-root-path/.well-known/openid-configuration',
+                },
+              },
+              { 'name': 'keycloak',
+                'icon': 'fa-key',
+                'token_key': 'access_token',
+                'remote_app': {
+                  'client_id': os.environ.get('OIDC_3A305E38C3B561F3_CLIENT_ID'),
+                  'client_secret': os.environ.get('OIDC_3A305E38C3B561F3_CLIENT_SECRET'),
+                  'client_kwargs': {
+                    'scope': 'openid'
+                  },
+                  'api_base_url': 'http://my.keycloak2.server//protocol/',
+                  'server_metadata_url': 'http://my.keycloak2.server//.well-known/openid-configuration',
+                },
+              }
+              ]
+              ").into())
+        ]), result);
     }
 }